+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.1.0 (Unreleased)
+
+IMPROVEMENTS:
+
+- remove provider version setting [GH-5](https://github.com/terraform-alicloud-modules/terraform-alicloud-eip/pull/5)
+
 ## 1.0.0 (January 10, 2020)
 
 - **Added:** `EIP module` [GH-1](https://github.com/terraform-alicloud-modules/terraform-alicloud-eip/pull/1)